externalize download content service
